+++ dot/network/discovery.go
@@ -1,3 +1,19 @@
+// Copyright 2019 ChainSafe Systems (ON) Corp.
+// This file is part of gossamer.
+//
+// The gossamer library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// The gossamer library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with the gossamer library. If not, see <http://www.gnu.org/licenses/>.
+
 package network
 
 import (
@@ -15,6 +31,7 @@ import (
 	"github.com/libp2p/go-libp2p-kad-dht/dual"
 )
 
+// discovery handles discovery of new peers via the kademlia DHT
 type discovery struct {
 	ctx                context.Context
 	dht                *dual.DHT
@@ -25,15 +42,15 @@ type discovery struct {
 	minPeers, maxPeers int
 }
 
-func newDiscovery(ctx context.Context, h libp2phost.Host, bootnodes []peer.AddrInfo, ds *badger.Datastore, pid protocol.ID) *discovery {
+func newDiscovery(ctx context.Context, h libp2phost.Host, bootnodes []peer.AddrInfo, ds *badger.Datastore, pid protocol.ID, min, max int) *discovery {
 	return &discovery{
 		ctx:       ctx,
 		h:         h,
 		bootnodes: bootnodes,
 		ds:        ds,
 		pid:       pid,
-		minPeers:  5,
-		maxPeers:  50, // TODO: pass this in
+		minPeers:  min,
+		maxPeers:  max,
 	}
 }
 
@@ -80,6 +97,10 @@ func (d *discovery) start() error {
 }
 
 func (d *discovery) stop() error {
+	if d.dht == nil {
+		return nil
+	}
+
 	return d.dht.Close()
 }
 
@@ -131,14 +152,19 @@ func (d *discovery) discoverAndAdvertise() error {
 			select {
 			case <-d.ctx.Done():
 				return
-			// case <-time.After(time.Minute):
-			// 	err := d.dht.Bootstrap(d.ctx)
-			// 	if err != nil {
-			// 		logger.Warn("failed to bootstrap DHT", "error", err)
-			// 		continue
-			// 	}
+			case <-time.After(time.Minute):
+				if len(d.h.Network().Peers()) > d.minPeers {
+					continue
+				}
 
-			// 	_ = d.dht.WAN.RefreshRoutingTable()
+				// reconnect to peers if peer count is low
+				for p := range peersToTry {
+					logger.Info("trying to connect to cached peer", "peer", p.ID)
+					err = d.h.Connect(d.ctx, *p)
+					if err != nil {
+						logger.Info("failed to connect to discovered peer", "peer", p.ID, "err", err)
+					}
+				}
 			case peer := <-peerCh:
 				if peer.ID == d.h.ID() || peer.ID == "" {
 					continue
@@ -155,16 +181,6 @@ func (d *discovery) discoverAndAdvertise() error {
 				} else {
 					d.h.Peerstore().AddAddrs(peer.ID, peer.Addrs, peerstore.PermanentAddrTTL)
 					peersToTry[&peer] = struct{}{}
-				}
-			}
-
-			if len(d.h.Network().Peers()) < d.minPeers {
-				for p := range peersToTry {
-					logger.Info("trying to connect to cached peer", "peer", p.ID)
-					err = d.h.Connect(d.ctx, *p)
-					if err != nil {
-						logger.Info("failed to connect to discovered peer", "peer", p.ID, "err", err)
-					}
 				}
 			}
 		}
