+++ dot/network/host_test.go
@@ -17,17 +17,12 @@
 package network
 
 import (
-	"context"
-	"fmt"
 	"net"
 	"testing"
 	"time"
 
 	"github.com/ChainSafe/gossamer/lib/common"
 	"github.com/ChainSafe/gossamer/lib/utils"
-	"github.com/libp2p/go-libp2p-core/peer"
-	"github.com/libp2p/go-libp2p-kad-dht/dual"
-	kbucket "github.com/libp2p/go-libp2p-kbucket"
 	ma "github.com/multiformats/go-multiaddr"
 
 	"github.com/stretchr/testify/require"
@@ -370,88 +365,4 @@ func TestStreamCloseMetadataCleanup(t *testing.T) {
 	// Verify that handshake data is cleared.
 	_, ok = info.getHandshakeData(nodeB.host.id(), true)
 	require.False(t, ok)
-}
-
-func createServiceHelper(t *testing.T, num int) []*Service {
-	t.Helper()
-	var srvcs []*Service
-	for i := 0; i < num; i++ {
-		config := &Config{
-			BasePath:    utils.NewTestBasePath(t, fmt.Sprintf("node%d", i)),
-			Port:        uint32(7001 + i),
-			RandSeed:    int64(1 + i),
-			NoBootstrap: true,
-			NoMDNS:      true,
-		}
-
-		srvc := createTestService(t, config)
-		srvc.noGossip = true
-		handler := newTestStreamHandler(testBlockAnnounceMessageDecoder)
-		srvc.host.registerStreamHandler("", handler.handleStream)
-
-		srvcs = append(srvcs, srvc)
-	}
-	return srvcs
-}
-
-// nolint
-func connectNoSync(t *testing.T, ctx context.Context, a, b *Service) {
-	t.Helper()
-
-	idB := b.host.h.ID()
-	addrB := b.host.h.Peerstore().Addrs(idB)
-	if len(addrB) == 0 {
-		t.Fatal("peers setup incorrectly: no local address")
-	}
-
-	a.host.h.Peerstore().AddAddrs(idB, addrB, time.Minute)
-	pi := peer.AddrInfo{ID: idB}
-
-	err := a.host.h.Connect(ctx, pi)
-	// retry connect if "failed to dial" error
-	if failedToDial(err) {
-		time.Sleep(TestBackoffTimeout)
-		err = a.host.h.Connect(ctx, pi)
-	}
-	require.NoError(t, err)
-}
-
-// nolint
-func wait(t *testing.T, ctx context.Context, a, b *dual.DHT) {
-	t.Helper()
-
-	// Loop until connection notification has been received.
-	// Under high load, this may not happen as immediately as we would like.
-	for a.LAN.RoutingTable().Find(b.LAN.PeerID()) == "" {
-		select {
-		case <-ctx.Done():
-			t.Fatal(ctx.Err())
-		case <-time.After(time.Millisecond * 5):
-		}
-	}
-}
-
-// Set `NoMDNS` to true and test routing via kademlia DHT service.
-func TestKadDHT(t *testing.T) {
-	if testing.Short() {
-		return
-	}
-
-	nodes := createServiceHelper(t, 3)
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
-	defer cancel()
-
-	_, err := nodes[2].host.dht.FindPeer(ctx, nodes[1].host.id())
-	require.Equal(t, err, kbucket.ErrLookupFailure)
-
-	connectNoSync(t, ctx, nodes[1], nodes[0])
-	connectNoSync(t, ctx, nodes[2], nodes[0])
-
-	// Can't use `connect` because b and c are only clients.
-	wait(t, ctx, nodes[1].host.dht, nodes[0].host.dht)
-	wait(t, ctx, nodes[2].host.dht, nodes[0].host.dht)
-
-	_, err = nodes[2].host.dht.FindPeer(ctx, nodes[1].host.id())
-	require.NoError(t, err)
 }
